modif calcul_suite.c et tableau.c et ajout tri.c

+++ td2/calcul_suite.c
@@ -3,12 +3,12 @@
 #include <time.h>
 #include <math.h>
 
-float racine(int x)
+double racine(int x)
 {	
-	float err=0.000001;
-	float deb=0.0;
-	float fin=x;
-	float res=(deb+fin)/2.0;
+	double err=0.000001;
+	double deb=0.0;
+	double fin=x;
+	double res=(deb+fin)/2.0;
 	
 	if((deb*deb<x+err)&&(deb*deb>x-err)) return deb;
 	if((fin*fin<x+err)&&(fin*fin>x-err)) return fin;
@@ -29,7 +29,7 @@ float racine(int x)
 	return res;
 }
 
-float calcul_suite(int nmax,int n)
+double calcul_suite(int nmax,int n)
 {	
 	if(n==nmax)
 	{
